+++ src/layouts/CommonLayout.tsx
@@ -1,5 +1,5 @@
-import Footer from "../components/Common/Footer";
-import Header from "../components/Common/Header";
+import Footer from '../components/Common/Footer';
+import Header from '../components/Common/Header';
 
 type Props = {
   children: React.ReactNode;
@@ -7,10 +7,10 @@ type Props = {
 
 export default function CommonLayout({ children }: Props) {
   return (
-    <div>
+    <>
       <Header />
       {children}
       <Footer />
-    </div>
+    </>
   );
 }
